requirements.txt
aiohttp==3.10.5
aiosignal==1.3.1
aniso8601==9.0.1
anyio==4.1.0
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
clickclick==20.10.2
colorama==0.4.6
connexion==2.14.2
cryptography==42.0.4
distlib==0.3.7
exceptiongroup==1.2.0
filelock==3.12.4
Flask==2.2.5
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
frozenlist==1.4.0
gevent==25.9.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.7
importlib-metadata==7.0.0
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jwt==1.3.1
loguru==0.7.0
lxml==4.9.3
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.1
path==16.7.1
Pillow==10.0.1
platformdirs==3.10.0
psutil==5.9.5
pycparser==2.21
pyrsistent==0.19.3
python-dotenv==1.0.1
python-multipart==0.0.7
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rpds-py==0.9.2
six==1.16.0
sniffio==1.3.0
starlette==0.36.2
swagger-ui-bundle==0.0.9
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
Unidecode==1.3.6
urllib3==2.0.7
virtualenv==20.24.5
Werkzeug==2.2.3
win32-setctime==1.1.0
xmltodict==0.13.0
yarl==1.9.2
zeep==4.2.1
zipp==3.17.0
psycopg2-binary==2.9.7
argon2-cffi==25.1.0
redis==5.0.7